# api.yaml
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/user/{user_id}/scenarios": {
      "get": {
        "operationId": "ScenariosController_getScenarios",
        "summary": "List all scenarios of a user",
        "description": "Returns all automated scenarios belonging to the specified user.",
        "parameters": [
          {
            "name": "user_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of scenarios fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Scenario"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden – user_id mismatch"
          }
        },
        "tags": [
          "scenarios"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "ScenariosController_create",
        "summary": "Create a new scenario",
        "description": "Creates a new automated scenario for the specified user.",
        "parameters": [
          {
            "name": "user_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateScenarioDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "New scenario added successfully"
          },
          "400": {
            "description": "Bad Request – validation failed"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden – user_id mismatch"
          }
        },
        "tags": [
          "scenarios"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/user/{user_id}/scenarios/{scenario_id}": {
      "get": {
        "operationId": "ScenariosController_getById",
        "summary": "Get a single scenario",
        "description": "Fetches a specific scenario owned by the user.",
        "parameters": [
          {
            "name": "user_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scenario_id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Scenario details retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Scenario"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden – user_id mismatch"
          },
          "404": {
            "description": "Scenario not found"
          }
        },
        "tags": [
          "scenarios"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "ScenariosController_update",
        "summary": "Update a scenario",
        "description": "Updates an existing scenario. Only the owner can modify it.",
        "parameters": [
          {
            "name": "user_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scenario_id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateScenarioDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Scenario updated successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden – user_id mismatch"
          },
          "404": {
            "description": "Scenario not found"
          }
        },
        "tags": [
          "scenarios"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "ScenariosController_delete",
        "summary": "Delete a scenario",
        "description": "Deletes a scenario. Only the owner can delete it.",
        "parameters": [
          {
            "name": "user_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scenario_id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Scenario deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden – user_id mismatch"
          },
          "404": {
            "description": "Scenario not found"
          }
        },
        "tags": [
          "scenarios"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/user/{user_id}/scenarios/{scenario_id}/state": {
      "put": {
        "operationId": "ScenariosController_setState",
        "summary": "Set scenario active state",
        "description": "Activates or deactivates the scenario. Use ?active=true to enable, ?active=false to disable.",
        "parameters": [
          {
            "name": "user_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scenario_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "active",
            "required": true,
            "in": "query",
            "example": true,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Scenario state updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Scenario"
                }
              }
            }
          },
          "400": {
            "description": "Missing or invalid active parameter"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden – user_id mismatch"
          },
          "404": {
            "description": "Scenario not found"
          }
        },
        "tags": [
          "scenarios"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "smart-home",
    "description": "Smart home application",
    "version": "1.0",
    "contact": {

    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "Scenario": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "scen-001",
            "description": "Unique identifier for Scenario"
          },
          "name": {
            "type": "string",
            "description": "Name of the scenario",
            "example": "Morning Routine"
          },
          "isActive": {
            "type": "boolean",
            "description": "Displays if scenario is active or not",
            "example": true
          },
          "actions": {
            "type": "array",
            "description": "Actions to perform in this scenario",
            "example": [
              {
                "deviceId": "550e8400-e29b-41d4-a716-446655440000",
                "lightIntensity": "bright"
              }
            ],
            "items": {
              "type": "object"
            }
          }
        },
        "required": [
          "id",
          "name",
          "isActive",
          "actions"
        ]
      },
      "CreateScenarioDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the scenario",
            "example": "Movie Night"
          },
          "actions": {
            "type": "array",
            "description": "Actions to perform in this scenario",
            "items": {
              "type": "object"
            },
            "example": [
              {
                "deviceId": "550e8400-e29b-41d4-a716-446655440000",
                "lightIntensity": "dim"
              }
            ]
          }
        },
        "required": [
          "name",
          "actions"
        ]
      },
      "UpdateScenarioDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Updated name of scenario",
            "example": "Morning Routine"
          },
          "actions": {
            "type": "array",
            "description": "Updated list of actions for this scenario",
            "items": {
              "type": "object"
            },
            "example": [
              {
                "deviceId": "550e8400-e29b-41d4-a716-446655440000",
                "lightIntensity": "dim"
              }
            ]
          }
        }
      }
    }
  }
}